Clean up patrol orders implementation
